services/runpod-inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
httpx==0.27.2
python-multipart==0.0.12
# PyTorch for real model weights (placeholder detectors work without it)
# torch==2.4.1+cpu --index-url https://download.pytorch.org/whl/cpu
# torchvision==0.19.1+cpu --index-url https://download.pytorch.org/whl/cpu
